+++ kits/firestore-translate-text/src/handlers.ts
@@ -55,15 +55,16 @@ export async function handleDocumentWrite(
   event: TranslateWriteEvent,
   ctx: HandlerContext
 ): Promise<void> {
-  if (!event.data) {
-    return;
-  }
-
   const { config, service } = ctx;
 
   logs.start(config);
   const context = toEventContext(event);
   await events.recordStartEvent({ change: event.data, context });
+
+  if (!event.data) {
+    await events.recordCompletionEvent({ params: event.params });
+    return;
+  }
 
   const { languages, inputFieldName, outputFieldName } = config;